impl ShmipcHandler for RegionServer
