+++ src/frontend/src/server.rs
@@ -80,7 +80,7 @@ impl Services {
                 None,
                 user_provider.clone(),
                 grpc_runtime,
-                Some(instance.clone()),
+                None,
             );
 
             result.push((Box::new(grpc_server), grpc_addr));
